+++ package.json
@@ -12,11 +12,15 @@
     "lint": "eslint --ext .vue,.ts,.js .",
     "lint:fix": "eslint --ext .vue,.ts,.js . --fix",
     "format": "prettier --write \"**/*.{js,ts,vue,css,scss,json,md}\"",
-    "lh": "yarn build && lhci autorun && node scripts/process-lighthouse-reports.js && node scripts/rename-lighthouse-reports.js && yarn lh:view",
-    "lh:view": "node scripts/open-lighthouse-report.js",
-    "lh:all": "node scripts/open-all-lighthouse-reports.js",
+    "lh": "yarn build && lhci autorun && yarn lh:process && yarn lh:rename && yarn lh:view",
+    "lh:process": "node scripts/process-lighthouse-reports.js",
+    "lh:rename": "node scripts/rename-lighthouse-reports.js",
+    "lh:view": "node scripts/open-lighthouse-reports.js 1",
+    "lh:view:recent": "node scripts/open-lighthouse-reports.js 5 7",
+    "lh:view:all": "node scripts/open-lighthouse-reports.js 12",
+    "lh:clean": "node scripts/cleanup-lighthouse-reports.js",
     "lh:cleanAll": "node scripts/cleanup-lighthouse-reports.js cleanAll",
-    "ci": "yarn build && node scripts/cleanup-lighthouse-reports.js",
+    "ci": "yarn build && yarn lh:clean && yarn lh && yarn lh:view",
     "fix": "yarn lint:fix && yarn format",
     "prepare": "husky install"
   },
